+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- "Create Variable" now works well from variable declarations
+
 ## [1.2.0] - 2020-04-04 - I handle async functions 🔮
 
 ### Added